Add files via upload
In this project, we wish to explore which factors impact flight delays and provide insights to airlines on which areas to focus their optimization efforts. This will also help prove the case for optimizing flights to avoid delays to ultimately save the airline industry a massive amount of money and help provide a better service experience for consumers.